fibonacci rabbit

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,18 +7,18 @@
 
 int main(void)
 {
-	unsigned long int f1, f2, fn;
+	float f1, f2, fn;
 	int c;
 
 	f1 = 1;
 	f2 = 2;
 
-	printf("%lu, %lu", f1, f2);
+	printf("%.0f, %.0f", f1, f2);
 
 	for (c = 3; c <= 98; c++)
 	{
 		fn = (f1 + f2);
-		printf(", %lu", fn);
+		printf(", %.0f", fn);
 
 		f1 = (f2);
 		f2 = (fn);
